Fix auto deply docs
