--- docs/source/analysis_techniques.rst
@@ -1,4 +1,0 @@
-Analysis Techniques
-===================
-
-This section consists of different analysis procedures that are used by MailingListParser for examining the author interactions from different mailing list threads or mails of the same thread.
