--- docs/source/driver_author_analysis.rst
@@ -1,7 +1,0 @@
-driver_author_analysis module
-=============================
-
-.. automodule:: driver_author_analysis
-    :members:
-    :undoc-members:
-    :show-inheritance:
