--- docs/source/driver_headers_imap.rst
@@ -1,7 +1,0 @@
-driver_headers_imap module
-==========================
-
-.. automodule:: driver_headers_imap
-    :members:
-    :undoc-members:
-    :show-inheritance:
