--- docs/source/driver_headers_mbox.rst
@@ -1,7 +1,0 @@
-driver_headers_mbox module
-==========================
-
-.. automodule:: driver_headers_mbox
-    :members:
-    :undoc-members:
-    :show-inheritance:
