--- docs/source/driver_thread_analysis.rst
@@ -1,7 +1,0 @@
-driver_thread_analysis module
-=============================
-
-.. automodule:: driver_thread_analysis
-    :members:
-    :undoc-members:
-    :show-inheritance:
