--- docs/source/modules.rst
@@ -1,13 +1,0 @@
-lib
-===
-
-.. toctree::
-   :maxdepth: 4
-
-   analysis
-   driver_author_analysis
-   driver_headers_imap
-   driver_headers_mbox
-   driver_thread_analysis
-   input
-   test_driver
