+++ requirements.txt
@@ -8,3 +8,4 @@ nltk>=3.2.2
 networkx>=1.11
 pytest>=3.0.5
 sphinx>=1.4.1
+plotly>=1.3.2
